+++ ceammc/CHANGELOG.md
@@ -11,6 +11,7 @@
   - proto.morse - morse encoder with specified speed and other settings
 - new object inlets:
   - metro.seq - second inlet added to set interval and third inlet to set pattern
+  - metro.pattern - second inlet added to set pattern
 - new object methods:
   - euclid method added to ui.radio to set euclid pattern when in checklist mode
   - hexbeat method added to ui.radio to set hexbeat pattern when in checklist mode